Changed shopManagerMenu to have simple responsibility Deleted Basket Renamed menu  package
